Rename package from hudder to main
