+++ src/main/java/io/github/ngspace/hudder/Hudder.java
@@ -6,10 +6,10 @@ import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.github.ngspace.hudder.hudder.HudCompilationManager;
-import io.github.ngspace.hudder.hudder.HudderRenderer;
-import io.github.ngspace.hudder.hudder.HudderTickEvent;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.HudCompilationManager;
+import io.github.ngspace.hudder.main.HudderRenderer;
+import io.github.ngspace.hudder.main.HudderTickEvent;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.utils.HudFileUtils;
 import io.github.ngspace.hudder.utils.testing.HudderUnitTestingCommand;
 import net.fabricmc.api.ModInitializer;
