+++ src/main/java/io/github/ngspace/hudder/compilers/EmptyCompiler.java
@@ -3,7 +3,7 @@ package io.github.ngspace.hudder.compilers;
 import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.compilers.utils.HudInformation;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 
 public class EmptyCompiler extends ATextCompiler {
 	@Override
