+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/AScriptingLanguageCompiler.java
@@ -14,8 +14,8 @@ import io.github.ngspace.hudder.data_management.BooleanData;
 import io.github.ngspace.hudder.data_management.NumberData;
 import io.github.ngspace.hudder.data_management.ObjectData;
 import io.github.ngspace.hudder.data_management.StringData;
-import io.github.ngspace.hudder.hudder.HudCompilationManager;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.HudCompilationManager;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.uielements.AUIElement;
 import io.github.ngspace.hudder.uielements.ItemElement;
 import io.github.ngspace.hudder.utils.HudFileUtils;
