+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/ATextCompiler.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import io.github.ngspace.hudder.Hudder;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.compilers.utils.HudInformation;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 
 public abstract class ATextCompiler {
