+++ src/main/java/io/github/ngspace/hudder/compilers/abstractions/AV2Compiler.java
@@ -11,8 +11,8 @@ import io.github.ngspace.hudder.compilers.utils.unifiedcomp.UnifiedCompiler.Bind
 import io.github.ngspace.hudder.compilers.utils.unifiedcomp.UnifiedCompiler.BindableFunction;
 import io.github.ngspace.hudder.compilers.utils.unifiedcomp.UnifiedCompiler.ConsumerBinder;
 import io.github.ngspace.hudder.compilers.utils.unifiedcomp.UnifiedCompiler.FunctionBinder;
-import io.github.ngspace.hudder.hudder.HudCompilationManager;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.HudCompilationManager;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
 import io.github.ngspace.hudder.v2runtime.functions.IV2Function;
 import io.github.ngspace.hudder.v2runtime.functions.V2FunctionHandler;
