+++ src/main/java/io/github/ngspace/hudder/main/HudCompilationManager.java
@@ -1,4 +1,4 @@
-package io.github.ngspace.hudder.hudder;
+package io.github.ngspace.hudder.main;
 
 import java.util.ArrayList;
 import java.util.List;
