+++ src/main/java/io/github/ngspace/hudder/main/config/ConfigMenu.java
@@ -1,4 +1,4 @@
-package io.github.ngspace.hudder.hudder.config;
+package io.github.ngspace.hudder.main.config;
 
 import java.io.File;
 import java.io.IOException;
