+++ src/main/java/io/github/ngspace/hudder/main/config/HudderConfig.java
@@ -1,4 +1,4 @@
-package io.github.ngspace.hudder.hudder.config;
+package io.github.ngspace.hudder.main.config;
 
 import java.io.File;
 import java.io.FileWriter;
