+++ src/main/java/io/github/ngspace/hudder/uielements/ColorVerticesElement.java
@@ -2,7 +2,7 @@ package io.github.ngspace.hudder.uielements;
 
 import com.mojang.blaze3d.vertex.VertexFormat;
 
-import io.github.ngspace.hudder.hudder.HudderRenderer;
+import io.github.ngspace.hudder.main.HudderRenderer;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.GuiGraphics;
