+++ src/main/java/io/github/ngspace/hudder/uielements/TextElement.java
@@ -1,6 +1,6 @@
 package io.github.ngspace.hudder.uielements;
 
-import io.github.ngspace.hudder.hudder.HudderRenderer;
+import io.github.ngspace.hudder.main.HudderRenderer;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.GuiGraphics;
