+++ src/main/java/io/github/ngspace/hudder/uielements/TextureElement.java
@@ -2,7 +2,7 @@ package io.github.ngspace.hudder.uielements;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 
-import io.github.ngspace.hudder.hudder.HudderRenderer;
+import io.github.ngspace.hudder.main.HudderRenderer;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.RenderType;
