+++ src/main/java/io/github/ngspace/hudder/uielements/TextureVerticesElement.java
@@ -3,7 +3,7 @@ package io.github.ngspace.hudder.uielements;
 import java.io.IOException;
 
 import io.github.ngspace.hudder.compilers.utils.CompileException;
-import io.github.ngspace.hudder.hudder.HudderRenderer;
+import io.github.ngspace.hudder.main.HudderRenderer;
 import io.github.ngspace.hudder.utils.HudFileUtils;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.GuiGraphics;
