+++ src/main/java/io/github/ngspace/hudder/utils/testing/HudderUnitTest.java
@@ -5,7 +5,7 @@ import java.util.HashMap;
 import io.github.ngspace.hudder.Hudder;
 import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
 import io.github.ngspace.hudder.compilers.abstractions.AV2Compiler;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
