+++ src/main/java/io/github/ngspace/hudder/v2runtime/methods/GUIMethods.java
@@ -3,7 +3,7 @@ package io.github.ngspace.hudder.v2runtime.methods;
 import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.compilers.utils.CompileState;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.uielements.GameHudElement;
 import io.github.ngspace.hudder.uielements.GameHudElement.GuiType;
 import io.github.ngspace.hudder.utils.ObjectWrapper;
