+++ src/main/java/io/github/ngspace/hudder/v2runtime/runtime_elements/WhileV2RuntimeElement.java
@@ -4,7 +4,7 @@ import io.github.ngspace.hudder.compilers.abstractions.ATextCompiler.CharPositio
 import io.github.ngspace.hudder.compilers.abstractions.AV2Compiler;
 import io.github.ngspace.hudder.compilers.utils.CompileException;
 import io.github.ngspace.hudder.compilers.utils.CompileState;
-import io.github.ngspace.hudder.hudder.config.HudderConfig;
+import io.github.ngspace.hudder.main.config.HudderConfig;
 import io.github.ngspace.hudder.v2runtime.V2Runtime;
 import io.github.ngspace.hudder.v2runtime.values.AV2Value;
